Project_Cengiz/requirements.txt
numpy
Pillow
matplotlib
tensorboard
tensorboardX
